* works now

+++ copypasta-manager.py
@@ -1,15 +1,22 @@
 #!/usr/bin/python
 # -*- coding: utf-8 -*-
 
-# Copyright 2010 Nils Dagsson Moskopp
+#       Copyright 2010 Nils Dagsson Moskopp // erlehmann
 
-# This program is free software. It comes without any warranty, to
-# the extent permitted by applicable law. You can redistribute it
-# and/or modify it under the terms of the Do What The Fuck You Want
-# To Public License, Version 2, as published by Sam Hocevar. See
-# http://sam.zoy.org/wtfpl/COPYING for more details.
-
-# see also 
+#       This program is free software; you can redistribute it and/or modify
+#       it under the terms of the GNU General Public License as published by
+#       the Free Software Foundation; either version 3 of the License, or
+#       (at your option) any later version.
+#       
+#       This program is distributed in the hope that it will be useful,
+#       but WITHOUT ANY WARRANTY; without even the implied warranty of
+#       MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#       GNU General Public License for more details.
+#       
+#       You should have received a copy of the GNU General Public License
+#       along with this program; if not, write to the Free Software
+#       Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
+#       MA 02110-1301, USA.
 
 from __future__ import with_statement
 
@@ -23,6 +30,8 @@ class CopypastaManager:
         builder.connect_signals(self)
 
         self.window = builder.get_object("window")
+
+        self.statusbar = builder.get_object("statusbar")
 
         self.pastaview = builder.get_object("textview_pasta")
 
@@ -46,6 +55,7 @@ class CopypastaManager:
 
         self.pastapath = os.path.join(os.getcwd(),"pastas")
         self.populate_pastatree(self.pastatree, self.pastapath)
+        self.print_status("file import", "Importiert: %s" % self.pastapath)
 
     def main(self):
         self.window.show_all()
@@ -59,8 +69,46 @@ class CopypastaManager:
         clipboard.set_text(text)
         clipboard.store()
 
+    def on_button_delete_clicked(self, button):
+        iter = self.pastaselection.get_selected()[1]
+        abspath = self.pastastore.get_value(iter, 2)
+
+        if os.path.isfile(abspath):
+            try:
+                os.remove(abspath)
+                self.pastastore.remove(iter)
+                self.print_status("file delete", "Gelöscht: %s" % abspath)
+            except OSError, e:
+                self.print_status("file delete", "Löschen fehlgeschlagen: %s" % e)
+        elif os.path.isdir(abspath):
+            self.print_status("file delete", "Löschen fehlgeschlagen: Ist ein Verzeichnis.")
+        
     def on_button_new_clicked(self, button):
-        self.pastastore.append(None, ["Neue Kopierpaste", "", ""])
+        iter = self.pastaselection.get_selected()[1]
+
+        try:
+            abspath = self.pastastore.get_value(iter, 2)
+        except TypeError:
+            abspath = self.pastapath
+
+        if os.path.isdir(abspath):
+            dir = abspath
+        elif os.path.isfile(abspath):
+            dir = os.path.dirname(abspath)
+            iter = None
+
+        title = "Neue Kopierpaste"
+        newfile = os.path.join(dir, title)
+
+        if not os.path.isfile(newfile):
+            try:
+                with open(newfile, 'w') as file:
+                    pass
+                self.pastastore.append(iter, [title, "", newfile])
+            except:
+                self.print_status("file creation", "Dateierstellung fehlgeschlagen: %s" % abspath)
+        else:
+            self.print_status("file creation", "Dateierstellung fehlgeschlagen: „%s“ existiert bereits." % title)
 
     def on_pastabuffer_changed(self, buffer):
         start, end = buffer.get_bounds()
@@ -70,24 +118,34 @@ class CopypastaManager:
         self.pastastore.set(iter, 1, text)
 
         abspath = self.pastastore.get_value(iter, 2)
-        with open(abspath, 'w') as file:
-            #file.seek(0)
-            file.write(text)
+        if os.path.isfile(abspath):
+            try:
+                with open(abspath, 'w') as file:
+                    file.write(text)
+            except IOError:
+                self.print_status("file open", "Aktualisierung fehlgeschlagen: %s" % abspath)
 
     def on_renderer_edited(self, cell, path, text, model):
         try:
             oldname = model[path][2]
             newname = os.path.join(os.path.dirname(model[path][2]),text)
-            os.rename(oldname, newname)
-            model[path][0] = text
-            model[path][2] = newname
-        except OSError:
-            pass
+
+            if not oldname == newname:
+                if not os.path.exists(newname):
+                    os.rename(oldname, newname)
+                    model[path][0] = text
+                    model[path][2] = newname
+                else:
+                    self.print_status("file rename", "Umbenennen fehlgeschlagen: Datei %s existiert." % newname)
+
+        except OSError, e:
+            self.print_status("file rename", "Umbenennen fehlgeschlagen: Wakarimasen lol.")
 
     def on_row_changed(self, treeselection):
         iter = treeselection.get_selected()[1]
-        content = self.pastastore.get_value(iter, 1)
-        self.pastabuffer.set_text(content)
+        if iter:
+            content = self.pastastore.get_value(iter, 1)
+            self.pastabuffer.set_text(content)
 
     def on_window_destroy(self, widget, data=None):
         gtk.main_quit()
@@ -102,13 +160,20 @@ class CopypastaManager:
         for filename in dirlist:
             abspath = os.path.join(path, filename)
 
-            try:                
-                with open(abspath) as file:
-                    content = file.read()
-                    self.pastastore.append(iter, [filename, content, abspath])
-            except IOError:
+            if os.path.isfile(abspath):
+                try:
+                    with open(abspath) as file:
+                        content = file.read()
+                        self.pastastore.append(iter, [filename, content, abspath])
+                except IOError:
+                    pass
+            if os.path.isdir(abspath):
                 parent = self.pastastore.append(iter, [filename, "", abspath])
                 self.populate_pastatree(treestore, abspath, parent)
+
+    def print_status(self, context, message):
+        cid = self.statusbar.get_context_id(context)
+        self.statusbar.push(cid, message)
 
 if __name__ == '__main__':
     cm = CopypastaManager()
